Added support for building message catalogs
The makefile is based on the one included in gettext ver. 0.18.2.
It has been modified to eliminate the autoconf dependency, but
the modifications require some GNU make extensions so other make
programs likely won't work. Because of this, and because gettext
isn't available everywhere, internationalization support is not
enabled by default.