Fixed character set and build of the man page.
Despite the XML file specifying UTF-8, xsltproc needs to be told
explicitly. It also builds the nroff file in the current directory by
default, which was supposed to have been moved in the last commit.
Also, added a link to the README.md file in the main HTML docs and
install the file.