Fixed some CI builds due to changes in the environments.
Don't turn the "'AutoPtr' is deprecated" compile warning into an error
on -Werror builds, since this is a problem that can only be addressed by
exiv2.  Switched a few builds to use slightly more modern distros.
Increase the git depth to avoid problems when skip ci is used.  Add a
Mac OS 10.14 build, and update Homebrew on the default builds (see
https://travis-ci.community/t/xcode-8-3-homebrew-outdated-error/3798).
Install the right package for the Docbook DTD in the FreeBSD builds.