Build the documentation in the dist target.
This includes the processed files in the tar ball and avoids the need
for xsltproc to build the package as long as the manual hasn't been
changed.  Added a distclean target to clean the targets that are built
during the dist rule. Plain clean shouldn't clean those (just AUTHORS
for now) because those can't be recreated on a tree coming from a dist
archive.