Add the --show-gpx option to create GPX output for images.
This is used to create a GPX file from a set of existing images that are
already GPS tagged. The resulting file can be used to correlate further
pictures, or can be used in any other application that uses GPX files
such as to map the path taken while taking the pictures.