Add support for writing image direction tags.
When the GPX file includes the course, it's used to write the GPSTrack
and GPSImgDirection EXIF tags.  This is controlled by three command-line
flags: --heading, which enables writing GPSTrack, --direction, which
enables writing GPSImgDirection and specifies the number of degrees of
difference between the heading and the camera direction, and
--max-heading, which gives the maximum change in degrees between track
points, beyond which no EXIF tag will be written.

TODO:
- find the right GPX tags to use
- maybe add a per-point magnetic or true north flag, set depending on
the tag used
- GUI support
- docs
- look through notes to see what I've missed